+++ NEWS.rst
@@ -157,7 +157,8 @@ image_align (:numref:`image_align`):
 
 cam_gen (:numref:`cam_gen`):
   * Added the option ``--camera-center``.
- 
+  * Can export an RPC camera model to .xml format (:numref:`cam_gen_rpc`).
+   
 dem_mosaic (:numref:`dem_mosaic`):
   * Added the option ``--weight-list`` for blending DEMs given external weights
     (:numref:`dem_mosaic_external_weights`).